refactor!: shrink feerates to u32 across the stack
A u32 sat/kvb caps at ~4.3M sat/vb-equivalent, three orders of
magnitude above any historical fee spike. Converts once at the bitcoind
RPC boundary and carries u32 through the feerate vote, the config's
feerate_base, and the client API; the fee math widens back to u64
before multiplying against vbytes.

+++ picomint-bitcoin-rpc/src/lib.rs
@@ -23,7 +23,7 @@ const MUTINYNET: &str = "000002855893a0a9b24eaffc5efc770558a326fee4fc10c9da22fc1
 
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
 pub struct Feerate {
-    pub sat_per_kvb: u64,
+    pub sat_per_kvb: u32,
 }
 
 /// Status of the Bitcoin RPC backend as reported by the monitor.
